+++ deploiement.bat
@@ -1,28 +1,44 @@
 @echo off
 
+:: Définir les variables
 set "lib=E:\Licence_3\Semestre-5\M.Niaina\SpringMVC\Framework\lib"
 set "SRC=E:\Licence_3\Semestre-5\M.Niaina\SpringMVC\Framework\src"
-set "JAR=E:\Licence_3\Semestre-5\M.Niaina\SpringMVC\Framework\jar"
+set "temp_src=E:\Licence_3\Semestre-5\M.Niaina\SpringMVC\Framework\temp_src"
+set "classes=E:\Licence_3\Semestre-5\M.Niaina\SpringMVC\Framework\classes"
+set "CLASSPATH=E:\Licence_3\Semestre-5\M.Niaina\SpringMVC\Test\lib"
 set "libParanamer=E:\Licence_3\Semestre-5\M.Niaina\SpringMVC\Framework\lib\paranamer-2.8.jar"
 set "libGson=E:\Licence_3\Semestre-5\M.Niaina\SpringMVC\Framework\lib\gson-2.8.2.jar"
-set "CLASSPATH=E:\Licence_3\Semestre-5\M.Niaina\SpringMVC\Test\lib"
-set "jarName=sprint13-2658" 
+set "jarName=sprint14-2658"
 
-if not exist "%JAR%" mkdir "%JAR%"
+:: Créer les répertoires temporaires si nécessaires
+if not exist "%temp_src%" mkdir "%temp_src%"
+if not exist "%classes%" mkdir "%classes%"
+
+:: Copier les fichiers .java dans temp_src
 cd "%SRC%"
-javac -parameters -cp "%lib%\*" -d "%JAR%/" *.java
+for /r %%F in (*.java) do (
+    copy "%%F" "%temp_src%" >nul
+)
+cd ..
 
-cd "%JAR%"
+:: Compiler les fichiers .java dans temp_src et générer les .class dans classes
+javac -parameters -cp "%lib%\*" -d "%classes%" "%temp_src%\*.java"
 
+:: Créer le fichier JAR à partir des .class
+cd "%classes%"
+jar -cvf "../%jarName%.jar" .
+cd ..
 
-jar -cvf "%jarName%.jar" .
+:: Copier le fichier JAR et les dépendances dans le CLASSPATH
+if not exist "%CLASSPATH%" mkdir "%CLASSPATH%"
+copy "%jarName%.jar" "%CLASSPATH%" >nul
+copy "%libParanamer%" "%CLASSPATH%" >nul
+copy "%libGson%" "%CLASSPATH%" >nul
 
+:: Nettoyer les répertoires temporaires
+rmdir /s /q "%temp_src%"
+rmdir /s /q "%classes%"
 
-copy   "%jarName%.jar" "%CLASSPATH%\"
-copy   "%libParanamer%" "%CLASSPATH%\"
-copy   "%libGson%" "%CLASSPATH%\"
-cd "%SRC%"
-rmdir /s /q "%JAR%"
-echo Jar créé avec succès.
-
+:: Fin
+echo Jar créé avec succès et copié dans %CLASSPATH%.
 pause
