+++ src/annotation/Daty.java
@@ -4,5 +4,5 @@ import java.lang.annotation.*;
 
 @Target(ElementType.FIELD)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface Date {
+public @interface Daty {
 }
