+++ src/controller/FrontController.java
@@ -1,11 +1,13 @@
 package controller;
 
-import jakarta.servlet.http.*;  
-import jakarta.servlet.*;  
+import jakarta.servlet.http.*;
+import jakarta.servlet.*;
 import java.io.*;
 import java.net.URLDecoder;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
+
 import annotation.*;
 import utility.Util;
 import utility.Mapping;
@@ -15,6 +17,7 @@ import jakarta.servlet.annotation.WebServlet;
 import java.lang.reflect.*;
 import com.google.gson.Gson;
 import jakarta.servlet.annotation.MultipartConfig;
+
 /**
  *
  * @author Pc
@@ -22,19 +25,20 @@ import jakarta.servlet.annotation.MultipartConfig;
 @MultipartConfig
 public class FrontController extends HttpServlet {
     HashMap<String, Mapping> myHashMap = new HashMap<>();
-    public void initVariable() throws Exception{
+
+    public void initVariable() throws Exception {
         try {
-             Util util = Util.getInstance();
+            Util util = Util.getInstance();
             myHashMap = util.getListControllerWithAnnotationMethodUrl("packageController", getServletConfig());
-    } catch (Exception e) {
+        } catch (Exception e) {
             throw e;
-    }
-    
+        }
+
     }
 
-    public void init() throws ServletException{
+    public void init() throws ServletException {
         try {
-            
+
             initVariable();
         } catch (Exception e) {
             e.printStackTrace();
@@ -46,52 +50,49 @@ public class FrontController extends HttpServlet {
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         PrintWriter out = response.getWriter();
-        try{
+        try {
             Util util = Util.getInstance();
-            //url apres nom de domaine
+            // url apres nom de domaine
             String requestURI = request.getRequestURI();
-            //chemin du context
+            // chemin du context
             String contextPath = request.getContextPath();
-            //url apres le context
+            // url apres le context
             String pathInfo = requestURI.substring(contextPath.length());
 
-           
-            
             int queryIndex = pathInfo.indexOf('?');
             String urlMapping = pathInfo;
             if (queryIndex != -1) {
                 urlMapping = pathInfo.substring(0, queryIndex);
             }
-            //util.isDuplicateUrlMapping(urlMapping ,"packageController", getServletConfig());
-            Mapping map = util.findMappingAssociateUrl(myHashMap,urlMapping);
-            
+            // util.isDuplicateUrlMapping(urlMapping ,"packageController",
+            // getServletConfig());
+            Mapping map = util.findMappingAssociateUrl(myHashMap, urlMapping);
+
             if (map == null) {
                 response.setContentType("text/html");
                 response.setStatus(HttpServletResponse.SC_NOT_FOUND);
-                out.write(Util.generateErreurHtml(response,"Pas mapping associe a cette URL"));
-                return ;
-            }else{
-                util.checkControllerContainsAttributMySession(map.getClassName(),request);
-                Object result = util.executeMethod(map,urlMapping,request,response);
+                out.write(Util.generateErreurHtml(response, "Pas mapping associe a cette URL"));
+                return;
+            } else {
+                Method methode = util.getMethode(map.getClassName(), map.getListMethodVerb(), urlMapping,
+                        request.getMethod());
 
-                if (result == null) {
-                    out.print("result null");
-                    throw new Exception("response null");
-                   
-                }
+                util.processValidationAndForward(methode, request, response);
 
-                if(util.isStringOrModelview(result)){
-                    if(result instanceof ModelView){
+                Object result = util.executeMethod(map, methode, request, response);
+
+                if (util.isStringOrModelview(result)) {
+                    if (result instanceof ModelView) {
                         ModelView modelview = (ModelView) result;
-                        util.redirectModelView(request,response,modelview);
-                    }else{
+                        util.redirectModelView(request, response, modelview);
+                    } else {
                         throw new Exception(" ERREUR 500 la methode est de type string => " + result);
                     }
-            }else{
-                throw new Exception(" ERREUR 500 la methode n'est pas de type string ou modelView " + result);
+                } else {
+                    throw new Exception(" ERREUR 500 la methode n'est pas de type string ou modelView " + result);
+                }
             }
-            }
-    }catch(Exception e){
+        } catch (Exception e) {
             e.printStackTrace();
             e.getStackTrace();
             e.getMessage();
@@ -99,11 +100,10 @@ public class FrontController extends HttpServlet {
             System.err.println(e.getMessage());
             response.setContentType("text/html");
             response.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
-            out.write(Util.generateErreurHtml(response,e.getMessage()));
-            
+            out.write(Util.generateErreurHtml(response, e.getMessage()));
+
         }
     }
-
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
@@ -111,7 +111,6 @@ public class FrontController extends HttpServlet {
         processRequest(request, response);
     }
 
- 
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
